Rewrite exec command on OS X. Fixes #1.
Sublime's normal exec routine runs commands in a login shell to ensure the
PATH from .bash_profile et al are respected. In this case, we explicitly do
want the supplied path rather than the shell path, so this patch rewrites
the kwargs to the exe routine to avoid being run in a login shell.